Create Bipartite_BFS.java
To check whether a given graph G=(V, E) is bipartite or not, we can use BFS algorithm where we try to color the graph G using two colors such that no two adjacent nodes share the same color. 

A graph is bipartite if you can split its vertices into two sets such that no edge connects nodes of the same set. 

A graph is bipartite if and only if it is 2-colorable or we can say a graph is bipartite if and only if it contains no cycles of odd length.